+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changelog (nionui)
 
 0.6.5 (UNRELEASED)
 ------------------
+- Add a declarative twist-down section widget.
 - Add Path object and method to add it to drawing context.
 - Add ability to bind text_color property of declarative label.